requirements.txt
# Core Framework
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Video Processing
opencv-python>=4.8.0
moviepy>=1.0.3
Pillow>=10.0.0

# AI Integration
google-generativeai>=0.3.0

# Caching & Storage
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
python-multipart>=0.0.6

# Development
watchdog>=3.0.0  # For auto-reload